+++ oci/optimizer_categories_data_source.go
@@ -1,0 +1,226 @@
+// Copyright (c) 2017, 2020, Oracle and/or its affiliates. All rights reserved.
+// Licensed under the Mozilla Public License v2.0
+
+package oci
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	oci_optimizer "github.com/oracle/oci-go-sdk/v27/optimizer"
+)
+
+func init() {
+	RegisterDatasource("oci_optimizer_categories", OptimizerCategoriesDataSource())
+}
+
+func OptimizerCategoriesDataSource() *schema.Resource {
+	return &schema.Resource{
+		Read: readOptimizerCategories,
+		Schema: map[string]*schema.Schema{
+			"filter": dataSourceFiltersSchema(),
+			"compartment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"compartment_id_in_subtree": {
+				Type:     schema.TypeBool,
+				Required: true,
+			},
+			"name": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"state": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"category_collection": {
+				Type:     schema.TypeList,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"items": {
+							Type:     schema.TypeList,
+							Computed: true,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									// Required
+
+									// Optional
+
+									// Computed
+									"compartment_id": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"description": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"estimated_cost_saving": {
+										Type:     schema.TypeFloat,
+										Computed: true,
+									},
+									"id": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"name": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"recommendation_counts": {
+										Type:     schema.TypeList,
+										Computed: true,
+										Elem: &schema.Resource{
+											Schema: map[string]*schema.Schema{
+												// Required
+
+												// Optional
+
+												// Computed
+												"count": {
+													Type:     schema.TypeInt,
+													Computed: true,
+												},
+												"importance": {
+													Type:     schema.TypeString,
+													Computed: true,
+												},
+											},
+										},
+									},
+									"resource_counts": {
+										Type:     schema.TypeList,
+										Computed: true,
+										Elem: &schema.Resource{
+											Schema: map[string]*schema.Schema{
+												// Required
+
+												// Optional
+
+												// Computed
+												"count": {
+													Type:     schema.TypeInt,
+													Computed: true,
+												},
+												"status": {
+													Type:     schema.TypeString,
+													Computed: true,
+												},
+											},
+										},
+									},
+									"state": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"time_created": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+									"time_updated": {
+										Type:     schema.TypeString,
+										Computed: true,
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+}
+
+func readOptimizerCategories(d *schema.ResourceData, m interface{}) error {
+	sync := &OptimizerCategoriesDataSourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).optimizerClient()
+
+	return ReadResource(sync)
+}
+
+type OptimizerCategoriesDataSourceCrud struct {
+	D      *schema.ResourceData
+	Client *oci_optimizer.OptimizerClient
+	Res    *oci_optimizer.ListCategoriesResponse
+}
+
+func (s *OptimizerCategoriesDataSourceCrud) VoidState() {
+	s.D.SetId("")
+}
+
+func (s *OptimizerCategoriesDataSourceCrud) Get() error {
+	request := oci_optimizer.ListCategoriesRequest{}
+
+	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
+		tmp := compartmentId.(string)
+		request.CompartmentId = &tmp
+	}
+
+	if compartmentIdInSubtree, ok := s.D.GetOkExists("compartment_id_in_subtree"); ok {
+		tmp := compartmentIdInSubtree.(bool)
+		request.CompartmentIdInSubtree = &tmp
+	}
+
+	if name, ok := s.D.GetOkExists("name"); ok {
+		tmp := name.(string)
+		request.Name = &tmp
+	}
+
+	if state, ok := s.D.GetOkExists("state"); ok {
+		request.LifecycleState = oci_optimizer.ListCategoriesLifecycleStateEnum(state.(string))
+	}
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(false, "optimizer")
+
+	response, err := s.Client.ListCategories(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response
+	request.Page = s.Res.OpcNextPage
+
+	for request.Page != nil {
+		listResponse, err := s.Client.ListCategories(context.Background(), request)
+		if err != nil {
+			return err
+		}
+
+		s.Res.Items = append(s.Res.Items, listResponse.Items...)
+		request.Page = listResponse.OpcNextPage
+	}
+
+	return nil
+}
+
+func (s *OptimizerCategoriesDataSourceCrud) SetData() error {
+	if s.Res == nil {
+		return nil
+	}
+
+	s.D.SetId(GenerateDataSourceHashID("OptimizerCategoriesDataSource-", OptimizerCategoriesDataSource(), s.D))
+	resources := []map[string]interface{}{}
+	category := map[string]interface{}{}
+
+	items := []interface{}{}
+	for _, item := range s.Res.Items {
+		items = append(items, CategorySummaryToMap(item))
+	}
+	category["items"] = items
+
+	if f, fOk := s.D.GetOkExists("filter"); fOk {
+		items = ApplyFiltersInCollection(f.(*schema.Set), items, OptimizerCategoriesDataSource().Schema["category_collection"].Elem.(*schema.Resource).Schema)
+		category["items"] = items
+	}
+
+	resources = append(resources, category)
+	if err := s.D.Set("category_collection", resources); err != nil {
+		return err
+	}
+
+	return nil
+}